[PATCH] Avoid usage of LoaderSymbol.
It can be dangerous. This patch is based on the previous one when r128
was in xf86-video-ati. The problem is that if gR128EntityIndex is not
explicitly marked as a public symbol, and weird things may happen when
compiling with hidden symbols.
Since the symbol is in the same shared binary of its callers, just use
it, instead of using LoaderSymbol. Could also just replace calls
getR128EntityIndex() to direct access to gR128EntityIndex.