Add EXA support
This introduces EXA acceleration which can improve performance and allow
2D acceleration to be used without XAA. Implemented hooks are Solid,
Copy and Composite. They appear to pass all rendercheck tests, except
the gradient test which XAA also fails. Tested on multiple color depths,
with and without DRI, with and without the composite extension. Hardware
cursor, Xvideo and page flipping are supported as well.
Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=47866

[ Michel Dänzer: Drop unrelated addition of #include "r128_version.h" ]

Signed-off-by: Connor Behan <connor.behan@gmail.com>
Signed-off-by: Michel Dänzer <michel.daenzer@amd.com>