Scale Xv with the vertical stretch ratio
If we are using a resolution lower than the native resolution of the
flat panel, the video overlay needs to know this. Otherwise a black bar
appears beneath the video. This is essentially Conn's patch from the bug
tracker except it doesn't always rely on info->PanelYRes. ACPI events
can change this without the driver's knowledge.
Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=6493

Signed-off-by: Connor Behan <connor.behan@gmail.com>