Update bitmasks for DDC
A RAGE128TR chipset with a VGA port uses different i2c clock bits than
the VGA cards previously tested. It seems reasonable to assume that
other Pro2 cards are set up this way as well. In case this is incorrect,
a newly added xf86I2CProbeAddress() should still allow a monitor to be
detected.

Signed-off-by: Connor Behan <connor.behan@gmail.com>
Tested-by: Tobias Powalowski <tobias.powalowski@googlemail.com>