Assume CRT in the absence of other monitors
This patch is still untested, but it sounds like a lot of people need
it. Assuming that there are no monitors when DDC fails (which causes X
to abort with UMS) does not make sense. Some people will always have DDC
fail. For example, the pins are not connected on an iMac G3. For another
example, some RAGE128TF cards are known to lie that there is no VGA port
in the connector table.

Signed-off-by: Connor Behan <connor.behan@gmail.com>